Advances in AI structures. Assets for AIBehaviourGraphs can now be saved with the scene on stand-alone builds. Modified Attacker and Attack so attacks can be spawned in various ways. Added Multishot attack. Advances in shaders. A ramp texture is now used to allow for smoother transitions inbetween shade cells.

+++ Game Project/Assets/_Game/Shaders/CustomDirectionalLight.shader
@@ -51,7 +51,7 @@
 				lightDir = normalize(lightDir);
 				float intensity = _CustomLightData.w;
 
-				half4 gbuffer2 = tex2D(_CameraGBufferTexture2, outVert.uv);
+				float4 gbuffer2 = tex2D(_CameraGBufferTexture2, outVert.uv);
